backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
passlib[bcrypt]
python-jose[cryptography]
supabase
python-dotenv
imap-tools
